refactor heatmap.go
